python3/requirements.txt
pandas
mysql.connector
mysql-connector-python
openpyxl
